Use VM Flavor in instance creation and counting
When provisioning different VM instances, different instances can have
different "flavors" where the architecture is the same, but things like
memory, disk etc. vary. Typically this also associates to a differing
maximum number of instances across VM flavors. When determining if a
particular VM instance can be created, the VM flavor's maximum number of
instances is compared to the current number of instances to ensure no
more than the maximum number of VMs are created.

However, when counting the current number of instances to determine if a
certain VM flavor instance can be created, only the instance tag (which
solely indicates the cluster) is used as a filter. If a VM of flavor A
with a lower maximum number of instances is ready to be provisioned and
several instances of flavor B are running on the same cluster, the ready
-to-be-provisioned flavor A VM might not get created (even
if no instances of flavor A are running) due to the number of flavor B
VMs being greater than flavor A's maximum number of instances.

A combination of the instance tag and VM flavor is now used to count the
number of Flavor X VMs running instead of the total VMs running on the
cluster.