Add a kustomize-config file to deployment build
The changes in the previous commit (40f23afd) used a configMapGenerator
to create ConfigMaps to mount the Task scripts to the TaskRun pod.
However, the generator appends a hash to the ConfigMap whenever its
updated. Since 'Task' is a custom resource, Kustomize doesn't know how
to auto-update the name. A kustomize-config file instructs Kustomize
where to look to append the new hash to the ConfigMap name every time
it's rebuilt.

KONFLUX-10713